refactor - remove shortcut to database files
as it was causing a circular dependancy
