--- backend/database/__init__.py
@@ -1,7 +1,0 @@
-from .initialize import setup as setup
-from .league import League as League
-from .event import Event as Event
-from .competitor import Competitor as Competitor
-from .result import Result as Result
-from .leagueResult import LeagueResult as LeagueResult
-from .search import Search as Search
