+++ backend/database/database.py
@@ -2,7 +2,7 @@ import os
 import psycopg2
 from typing import Any, Tuple
 
-DATABASE_URL = os.environ["DATABASE_URL"]
+DATABASE_URL = os.environ.get("DATABASE_URL", "")
 
 
 class DatabaseConnection:
